added fix (#21)
* added fix

* refactored the code

* added some changes

* removed comment
