+++ src/cli.ts
@@ -16,8 +16,7 @@ import { getSignature, sendToForDefi } from './forDefi/forDefi'
 import { createWithdrawalTransaction, sendSignedWithdrawalTransaction } from './forDefi/withdrawal'
 import { log, logError, logInfo, logSuccess, logWarning } from './output'
 import { colorCodes } from "./constants"
-import { fetchMirrorFunds } from "./mirrorFunds/main"
-import { getRpcUrl, submitForDefiTxn } from './flareContract'
+import { getRpcUrl, submitForDefiTxn, fetchMirrorFunds } from './flareContract'
 import { contractTransactionName } from './flareContractConstants'
 import fs from 'fs'
 import { forDefiDirectory, forDefiUnsignedTxnDirectory } from './constants'
