+++ src/flareContract.ts
@@ -1,4 +1,4 @@
-import { ClaimRewardsInterface, RegisterAddressInterface, UnsignedTxJson } from "./interfaces";
+import { ClaimRewardsInterface, Context, RegisterAddressInterface, UnsignedTxJson } from "./interfaces";
 import { SignatureRequest } from "@flarenetwork/flarejs/dist/common"
 import { bech32 } from "bech32";
 import { BigNumber, ethers } from "ethersV5";
@@ -8,11 +8,17 @@ import { colorCodes, getConfig, forDefiDirectory, forDefiUnsignedTxnDirectory } 
 import { NetworkConfig } from "./config";
 import {
   getAddressBinderABI, getFlareContractRegistryABI, defaultContractAddresses, addressBinderContractName,
-  validatorRewardManagerContractName, contractTransactionName, getValidatorRewardManagerABI
+  validatorRewardManagerContractName, contractTransactionName, getValidatorRewardManagerABI, pChainStakeMirror, getPChainStakeMirrorABI
 } from "./flareContractConstants";
-import { prefix0x, saveUnsignedTxJson } from "./utils";
+import { prefix0x, saveUnsignedTxJson, integerToDecimal } from "./utils";
 import { walletConstants } from "./screenConstants";
 import { exit } from "process";
+
+type DelegatedAmount = {
+  stakeAmount: number,
+  startTime: Date,
+  endTime: Date
+}
 
 /**
  * @description checks if the address is registered with the addressBinder contract
@@ -277,4 +283,82 @@ async function signContractTransaction(wallet: keyof typeof walletConstants, uns
     const signedTx = await wallet.signTransaction(unsignedTx)
     await contract.provider.sendTransaction(signedTx)
   }
+}
+
+////////// MIRROR FUND /////////
+// fetches current validator info
+const fetchValidatorInfo = async (ctx: Context) => {
+  const validator = await ctx.pchain.getCurrentValidators();
+  return validator;
+};
+
+// fetches pending validator info
+const fetchPendingValidatorInfo = async (ctx: Context) => {
+  const pendingValidator = await ctx.pchain.getPendingValidators();
+  return pendingValidator;
+};
+
+// fetches the delegation stake (from both current and pending validator) for the current user
+const fetchDelegateStake = async (ctx: Context, validatorFunction: (ctx: Context) => {}) => {
+  const validatorsInfo = await validatorFunction(ctx);
+  const validatorData = (validatorsInfo as any)?.validators;
+  let userStake = [];
+  for (let i = 0; i < validatorData.length; i++) {
+    for (let j = 0; j < (validatorData[i].delegators && validatorData[i].delegators?.length); j++) {
+      if (
+        validatorData[i].delegators[j] &&
+        validatorData[i].delegators[j].rewardOwner.addresses.includes(ctx.pAddressBech32)
+
+      ) {
+        const startDate = new Date(parseInt(validatorData[i]?.delegators[j]?.startTime) * 1000)
+        const endDate = new Date(parseInt(validatorData[i]?.delegators[j]?.endTime) * 1000)
+        userStake.push({
+          stakeAmount: parseFloat(validatorData[i]?.delegators[j]?.stakeAmount) / 1e9,
+          startTime: startDate,
+          endTime: endDate
+        });
+      }
+    }
+  }
+  return userStake;
+};
+
+// calculates the total amount of delegation
+const getTotalFromDelegation = (data: DelegatedAmount[]) => {
+  let total = 0;
+  for (let i = 0; i < data.length; i++) {
+    total += data[i].stakeAmount;
+  }
+  return total;
+};
+
+/**
+ * @description returns the mirror fund details
+ * @param ctx - context
+ * @returns - total mirror funds and funds with start and end time
+ */
+export async function fetchMirrorFunds(ctx: Context) {
+  console.log("Checking your Mirror Funds ...")
+  // fetch from the contract
+  const rpcUrl = getRpcUrl(ctx.config.hrp);
+  const pChainStakeMirrorContractAddress = await getContractAddress(ctx.config.hrp, pChainStakeMirror)
+  const provider = new ethers.providers.JsonRpcProvider(rpcUrl);
+  const abi = getPChainStakeMirrorABI() as ethers.ContractInterface
+  const contract = new ethers.Contract(pChainStakeMirrorContractAddress, abi, provider);
+  const stakedAmount = await contract.balanceOf(ctx.cAddressHex);
+  const stakedAmountInFLR = parseFloat(integerToDecimal(stakedAmount.toString(), 18));
+  // fetch for the chain
+  const delegationToCurrentValidator = await fetchDelegateStake(ctx, fetchValidatorInfo);
+  const delegationToPendingValidator = await fetchDelegateStake(ctx, fetchPendingValidatorInfo);
+  const totalDelegatedAmount =
+    getTotalFromDelegation(delegationToCurrentValidator) +
+    getTotalFromDelegation(delegationToPendingValidator)
+  const totalInFLR = parseFloat(totalDelegatedAmount.toString());
+  return {
+    "Total Mirrored Amount": `${totalInFLR} FLR`,
+    "Mirror Funds Details": {
+      ...delegationToCurrentValidator,
+      ...delegationToPendingValidator
+    }
+  };
 }
